Module ocsf::dictionary

Structs

DictAttribute

DictType

Enums

TypeNames

Constants

ACCESSED_TIME

Accessed Time - The time when the file was last accessed.

ACCESSOR

Accessor - The name of the user who last accessed the object.

ACCESS_LIST

Access List - The list of requested access rights.

ACCESS_MASK

Access Mask - The access mask in a platform-native format.

ACCESS_RESULT

Access Check Result - The list of access check results.

ACCOUNT_NAME

Account Name - The name of the account (e.g. AWS Account Name).

ACCOUNT_TYPE

Account Type - The user account type, normalized to the caption of ‘account_type_id’. In the case of ‘Other’, it is defined by the event source.

ACCOUNT_TYPE_ID

Account Type ID - The normalized user account type identifier.

ACCOUNT_UID

Account UID - The unique identifier of the account (e.g. AWS Account ID).

ACTIVITY_ID

Activity ID - The normalized identifier of the activity that triggered the event.

ACTIVITY_NAME

Activity - The event activity name, as defined by the activity_id.

ACTOR

Actor - The actor object describes details about the user/role/process that was the source of the activity.

ACTUAL_PERMISSIONS

Actual Permissions - The permissions that were granted to the in a platform-native format.

ALERT

Client TLS Alert - The integer value of TLS alert if present. The alerts are defined in the TLS specification in RFC-2246.

ALGORITHM

Algorithm - The hash algorithm used to create the digital fingerprint, normalized to the caption of ‘algorithm_id’. In the case of ‘Other’, it is defined by the event source.

ALGORITHM_ID

Algorithm ID - The identifier of the normalized hash algorithm, which was used to create the digital fingerprint.

ANSWERS

DNS Answer - The Domain Name System (DNS) answers.

API

API details - API object details information pertaining to the API calls

APP_NAME

Application Name - The name of the application that is associated with the event or object.

ARCHITECTURE

Architecture - Architecture is a shorthand name describing the type of computer hardware the packaged software is meant to run on.

ARGS

HTTP Arguments - The arguments sent along with the HTTP request.

ATTACKS

Attacks - An array of attacks associated with an event.

ATTEMPT

Attempt - The delivery attempt.

ATTRIBUTES

Attributes - The bitmask value that represents the file attributes.

AUTHORIZATIONS

Authorization Information - This object provides details such as authorization outcome, associated policies related to activity/event.

AUTH_PROTOCOL

Auth Protocol - The authentication protocol as defined by the caption of ‘auth_protocol_id’. In the case of ‘Other’, it is defined by the event source.

AUTH_PROTOCOL_ID

Auth Protocol ID - The normalized identifier of the authentication protocol used to create the user session.

AUTH_TYPE

Authentication Type - The agreed upon authentication type, normalized to the caption of ‘auth_type_id’. In the case of ‘Other’, it is defined by the event source.

AUTH_TYPE_ID

Authentication Type ID - The normalized identifier of the agreed upon authentication type. See specific usage.

AUTOSCALE_UID

Autoscale UID - The unique identifier of the cloud autoscale configuration.

BANNER

SMTP Banner - The initial SMTP connection response that a messaging server receives after it connects to a email server.

BASE_ADDRESS

Base Address - The memory address where the module was loaded.

BASE_SCORE

Base Score - The base score as reported by the event source. See specific usage.

BIG_ENDIAN

Big Endian - A boolean indicating whether the most significant byte is stored/transmitted first.

BIOS_DATE

BIOS Date - The BIOS date. For example: 03/31/16.

BIOS_MANUFACTURER

BIOS Manufacturer - The BIOS manufacturer. For example: LENOVO.

BIOS_VER

BIOS Version - The BIOS version. For example: LENOVO G5ETA2WW (2.62).

BOUNDARY

Boundary - The boundary of the connection, normalized to the caption of ‘boundary_id’. In the case of ‘Other’, it is defined by the event source. For cloud connections, this translates to the traffic-boundary(same VPC, through IGW, etc.). For traditional networks, this is described as Local, Internal, or External.

BOUNDARY_ID

Boundary ID - The normalized identifier of the boundary of the connection. For cloud connections, this translates to the traffic-boundary (same VPC, through IGW, etc.). For traditional networks, this is described as Local, Internal, or External.

BUILD

OS Build - The operating system build number.

BYTES

Total Bytes - The total number of bytes (in and out).

BYTES_IN

Bytes In - The number of bytes sent from the destination to the source.

BYTES_OUT

Bytes Out - The number of bytes sent from the source to the destination.

CAPABILITIES

Capabilities - A list of RDP capabilities.

CAPTION

Caption - A short description or caption of the device. For example: Scanner 1 or Database Manager.

CATEGORIES

Website Categorization - The Website categorization names, as defined by category_ids enum values.

CATEGORY

Category - The object category. See specific usage.

CATEGORY_IDS

Website Categorization IDs - The Website categorization identifies.

CATEGORY_NAME

Category - The event category name, as defined by category_uid value.

CATEGORY_UID

Category ID - The category unique identifier of the event.

CC

Cc - The email header Cc values, as defined by RFC 5322.

CERTIFICATE

Certificate - The certificate object containing information about the digital certificate.

CERTIFICATE_CHAIN

Certificate Chain - The Chain of Certificate Serial Numbers field provides a chain of Certificate Issuer Serial Numbers leading to the Root Certificate Issuer.

CHALLENGE

Challenge - The VNC challenge

CHALLENGE_RESPONSE

Challenge Response - The VNC challenge response

CHASSIS

Chassis - The chassis type describes the system enclosure or physical form factor. Such as the following examples for Windows Windows Chassis Types

CIPHER

Cipher Suite - The negotiated cipher suite.

CIS_BENCHMARK_RESULT

CIS Benchmark Result - The CIS benchmark result.

CITY

City - The name of the city.

CLASS

Class - The class name of the object. See specific usage.

CLASSIFICATIONS

Classifications - The list of malware classifications, normalized to the captions of the classifcation_id values. In the case of ‘Other’, they are defined by the event source.

CLASSIFICATION_IDS

Classification IDs - The list of normalzied identifiers of the malware classifications. Reference: STIX Malware Types

CLASS_NAME

Class - The event class name, as defined by class_uid value.

CLASS_UID

Class ID - The unique identifier of a class. A Class describes the attributes available in an event.

CLIENT_CIPHERS

Client Cipher Suites - The client cipher suites that were exchanged during the TLS handshake negotiation.

CLIENT_DIALECTS

Client Dialects - The list of SMB dialects that the client speaks.

CLIENT_HASSH

Client HASSH - The Client HASSH fingerprinting object.

CLOUD

Cloud - Describes details about the Cloud enviroment where the event was originally created or logged.

CLOUD_PARTITION

Cloud Partition - The canonical cloud partition name to which the region is assigned (e.g. AWS Partitions: aws, aws-cn, aws-us-gov).

CMD_LINE

Command Line - The full command line used to launch an application, service, process, or job. For example: ssh user@10.0.0.10. If the command line is unavailable or missing, the empty string '' is to be used

CODE

Response Code - The numeric response sent to a request.

CODES

Response Codes - The list of numeric responses sent to a request.

COLOR_DEPTH

Color Depth - The numeric color depth.

COLOR_MAX

Color Maximum - The maximum color value (with ‘n’ bits this would result in a 2^n-1 maximum value).

COLOR_SHIFT

Color Shift - The color shift value represents the number of shifts needed in order to get the color value in a pixel to the least significant bit.

COMMAND

Command - The command name.

COMMAND_RESPONSE

Command Response - The response to the command.

COMMAND_RESPONSES

Command Responses - The responses to the command.

COMMENT

Comment - The user-provided comment.

COMPANY_NAME

Company Name - The name of the company that published the file. For example: Microsoft Corporation.

COMPLIANCE

Compliance - The complaince object provides context to compliance findings (e.g., a check against a specific regulatory or best practice framework such as CIS or NIST) and contains compliance related details.

COMPONENT

Component - The name or relative pathname of a sub-component of the data object, if applicable. For example: attachment.doc, attachment.zip/bad.doc, or part.mime/part.cab/part.uue/part.doc.

CONFIDENCE

Confidence - The confidence of the reported event severity as a percentage: 0%-100%.

CONFIDENTIALITY

Confidentiality - The file content confidentiality, normalized to the confidentiality_id value. In the case of ‘Other’, it is defined by the event source.

CONFIDENTIALITY_ID

Confidentiality ID - The normalized identifier of the file content confidentiality indicator.

CONNECTION_INFO

Connection Info - The network connection information.

CONNECTION_UID

Connection Identifier - The network connection identifier.

CONTAINER

Container - The container information.

CONTENT_TYPE

HTTP Content Type - The request header that identifies the original media type of the resource (prior to any content encoding applied for sending).

CONTINENT

Continent - The name of the continent.

COORDINATES

Coordinates - A two-element array, containing a longitude/latitude pair. The format conforms with GeoJSON. For example: [-73.983, 40.719].

CORRELATION_UID

Correlation UID - The unique identifier used to correlate events.

COUNT

Count - The number of times that events in the same logical group occurred during the event Start Time to End Time period.

COUNTRY

Country - The ISO 3166-1 Alpha-2 country code. For the complete list of country codes see ISO 3166-1 alpha-2 codes.Note: The two letter country code should be capitalized. For example: US or CA.

CPU_BITS

CPU Bits - The cpu architecture, the number of bits used for addressing in memory. For example: 32 or 64.

CPU_CORES

CPU Cores - The number of processor cores in all installed processors. For Example: 42.

CPU_COUNT

CPU Count - The number of physical processors on a system. For example: 1.

CPU_SPEED

Processor Type - The speed of the processor in Mhz. For Example: 4200.

CPU_TYPE

Processor Type - The processor type. For example: x86 Family 6 Model 37 Stepping 5.

CREATED_TIME

Created Time - The time when the object was created. See specific usage.

CREATE_MASK

Create Mask - The original Windows mask that is required to create the object.

CREATOR

Creator - The user that created the object associated with event. See specific usage.

CREDENTIAL_UID

User Credential ID - The unique identifier of the user’s credential. For example, AWS Access Key ID.

CRITICALITY

Criticality - Criticality of a resource/object in question

CUSTOMER_UID

Customer UID - The unique customer identifier.

CVE

CVE - The Common Vulnerabilities and Exposures (CVE).

CVES

CVE List - List of Common Vulnerabilities and Exposures (CVE).

CVSS

CVSS Score - The CVSS object details Common Vulnerability Scoring System (CVSS) scores from the advisory that are related to the vulnerability.

CWE_UID

CWE UID - The Common Weakness Enumeration (CWE) unique identifier. For example: CWE-787.

CWE_URL

CWE URL - Common Weakness Enumiration (CWE) definition URL. For example: https://cwe.mitre.org/data/definitions/787.html.

DATA

Data - The additional data that is associated with the event or object. See specific usage.

DATA_BUCKET

Data Bucket - The name of the data bucket (e.g. bucket name for AWS/GCP and blob name for Azure).

DCE_RPC

Distributed Computing Environment/Remote Procedure Call (DCE/RPC) - The DCE/RPC object describes the remote procedure call system for distributed computing environments.

DECISION

Authorization Decision/Outcome - Decision/outcome of the authorization mechanism (e.g. Approved, Denied)

DELIVERED_TO

Delivered To - The Delivered-To email header field.

DEPTH

CVSS Depth - The CVSS depth represents a depth of the equation used to calculate CVSS score.

DESC

Description - The description that pertains to the object or event. See specific usage.

DESKTOP_DISPLAY

Desktop Display - The desktop display affiliated with the event

DETAILS

Details - Details of an entity. See specific usage

DETECTION_UID

Detection UID - The associated unique detection event identifier. For example: detection response events include the Detection ID of the original event.

DEVELOPER_UID

Developer UID - The developer ID on the certificate that signed the file.

DEVICE

Device - The device that reported the event.

DEVICES

Devices - The object describes details related to the list of devices.

DIALECT

Dialect - The negotiated protocol dialect.

DIRECTION

Direction - The direction of the initiated connection, traffic, or email, normalized to the caption of the direction_id value. In the case of ‘Other’, it is defined by the event source.

DIRECTION_ID

Direction ID - The normalized identifier of the direction of the initiated connection, traffic, or email.

DISPOSITION

Disposition - The event disposition name, normalized to the caption of the disposition_id value. In the case of ‘Other’, it is defined by the event source.

DISPOSITION_ID

Disposition ID - When security issues, such as malware or policy violations, are detected and possibly corrected, then disposition_id describes the action taken by the security product.

DKIM

DKIM Status - The DomainKeys Identified Mail (DKIM) status of the email.

DKIM_DOMAIN

DKIM Domain - The DomainKeys Identified Mail (DKIM) signing domain of the email.

DKIM_SIGNATURE

DKIM Signature - The DomainKeys Identified Mail (DKIM) signature used by the sending/receiving system.

DMARC

DMARC Status - The Domain-based Message Authentication, Reporting and Conformance (DMARC) status of the email.

DMARC_OVERRIDE

DMARC Override - The Domain-based Message Authentication, Reporting and Conformance (DMARC) override action.

DMARC_POLICY

DMARC Policy - The Domain-based Message Authentication, Reporting and Conformance (DMARC) policy status.

DOMAIN

Domain - The name of the domain.

DOMAIN_INFO

Domain Information - The registration information pertaining to a domain.

DRIVER

Kernel Driver - The driver that was loaded/unloaded into the kernel

DST_ENDPOINT

Destination Endpoint - The network destination endpoint.

DURATION

Duration - The event duration or aggregate time, the amount of time the event covers from start_time to end_time in milliseconds.

EDITION

OS Edition - The operating system edition. For example: Professional.

EMAIL

Email - The email object.

EMAIL_ADDR

Email Address - The user’s email address.

EMAIL_AUTH

Email Authentication - The SPF, DKIM and DMARC attributes of an email.

END_TIME

End Time - The end time of a time period. See specific usage.

ENRICHMENTS

Enrichments - The additional information from an external data source, which is associated with the event. For example add location information for the IP address in the DNS answers: [{"name": "answers.ip", "value": "92.24.47.250", "type": "location", "data": {"city": "Socotra", "continent": "Asia", "coordinates": [-25.4153, 17.0743], "country": "YE", "desc": "Yemen"}}]

ENTITY

Entity - The managed entity that is being acted upon.

ENTITY_RESULT

Entity Result - The updated managed entity.

EPOCH

Epoch - The software package epoch. Epoch is a way to define weighted dependencies based on version numbers.

ERROR

Error Code - Error Code

ERROR_MESSAGE

Error Message - Error Message

EXIT_CODE

Exit Code - The exit code reported by a process when it terminates. The convention is that zero indicates success and any non-zero exit code indicates that some error occurred.

EXPIRATION_TIME

Expiration Time - The expiration time. See specific usage.

EXPOSED_PORT

Port - The IP port number exposed by container. For example 0.0.0.0:49155-> <<exposed_port>>/tcp

EXTENSION_LIST

Extension List - The list of TLS extensions

FEATURE

Feature - The feature that reported the event.

FILE

File - The file that pertains to the event or object. See specific usage.

FILE_DIFF

File Diff - File content differences used for change detection. For example, a common use case is to identify itemized changes within INI or configuration/property setting values.

FILE_RESULT

File Result - The result of the file change. It should contain the new values of the changed attributes.

FINDING

Finding - Finding object provides details related to a finding generated by security tool

FINGERPRINT

Fingerprint - The digital fingerprint associated with an object.

FINGERPRINTS

Fingerprints - An array of digital fingerprint objects.

FIRST_SEEN_TIME

First Seen - The initial detection time of the activity or object. See specific usage

FLAGS

Flags - The list of communication flags, normalized to the captions of the flag_ids values. In the case of ‘Other’, they are defined by the event source.

FLAG_IDS

Communication Flag IDs - The list of normalized identifiers of the communication flag IDs.

FOLDER

Folder - The folder that pertains to the event.

FROM

From - The email header From values, as defined by RFC 5322.

FUNCTION_KEYS

Function Keys - The number of function keys on client keyboard.

FUNCTION_NAME

Function Name - The entry-point function of the module. The system calls the entry-point function whenever a process or thread loads or unloads the module.

GROUP

Group - The group object associated with an entity such as user, policy, or rule.

GROUPS

Groups - The groups to which an entity belongs. See specific usage.

GROUP_NAME

Group Name - The name of the group that the resource belongs to.

HANDSHAKE_DUR

Handshake Duration - The amount of total time for the TLS handshake to complete after the TCP connection is established, including client-side delays, in milliseconds.

HOSTNAME

Hostname - The hostname of an endpoint or a device.

HTTP_COOKIES

HTTP Cookies - The cookies object describes details about HTTP cookies

HTTP_HEADERS

HTTP Headers - Additional HTTP headers of an HTTP request or response.

HTTP_METHOD

HTTP Method - The HTTP request method indicates the desired action to be performed for a given resource. Expected values:

TRACE

CONNECT

OPTIONS

HEAD

DELETE

POST

PUT

GET

HTTP_ONLY

HTTP Only - A cookie attribute to make it inaccessible via JavaScript

HTTP_REQUEST

HTTP Request - The HTTP Request made to a web server.

HTTP_RESPONSE

HTTP Response - The HTTP Response from a web server to a requester.

HTTP_STATUS

HTTP Status - The Hypertext Transfer Protocol (HTTP) status code returned to the client.

HW_INFO

Hardware Info - The device hardware information.

HYPERVISOR

Hypervisor - The name of the hypervisor running on the device. For example, Xen, VMware, Hyper-V, VirtualBox, etc.

IDENTIFIER_COOKIE

Identifier Cookie - The client identifier cookie during client/server exchange.

IDP

Identity Provider - This object describes details about the Identity Provider used.

IMAGE

Image - The image used as a template to run a container or virtual machine.

IME

IME - The Input Method Editor (IME) file name.

IMEI

IMEI - The International Mobile Station Equipment Identifier that is associated with the device.

INJECTION_TYPE

Injection Type - The process injection method, normalized to the caption of the injection_type_id value. In the case of ‘Other’, it is defined by the event source.

INJECTION_TYPE_ID

Injection Type ID - The normalized identifier of the process injection method.

INSTANCE_UID

Instance ID - The unique identifier of a VM instance.

INTEGRITY

Integrity - The process integrity level, normalized to the caption of the direction_id value. In the case of ‘Other’, it is defined by the event source (Windows only).

INTEGRITY_ID

Integrity Level - The normalized identifier of the process integrity level (Windows only).

INTERFACE_NAME

Network Interface Name - The name of the network interface (e.g. eth2).

INTERFACE_UID

Network Interface ID - The unique identifier of the network interface.

INTERMEDIATE_IPS

Intermediate IP Addresses - The intermediate IP Addresses. For example, the IP addresses in the HTTP X-Forwarded-For header.

INVOKED_BY

Invoked by - The name of the service that invoked the activity as described in the event.

IP

IP Address - The IP address, in either IPv4 or IPv6 format.

ISP

ISP - The name of the Internet Service Provider (ISP).

ISSUER

Issuer Details - The identifier of the session issuer.

ISSUER_DN

Issuer Distinguished Name - The certificate issuer distinguished name.

ISSUER_NAME

Issuer Name - The certificate issuer name.

IS_CLEARTEXT

Cleartext Credentials - Indicates whether the credentials were passed in clear text.Note: True if the credentials were passed in a clear text protocol such as FTP or TELNET, or if Windows detected that a user’s logon password was passed to the authentication package in clear text.

IS_COMPLIANT

Compliant Device - The event occurred on a compliant device.

IS_DEFAULT

Default Value - The indication of whether the value is from a default value name. For example, the value name could be missing.

IS_MANAGED

Managed Device - The event occurred on a managed device.

IS_ON_PREMISES

On Premises - The indication of whether the location is on premises.

IS_PERSONAL

Personal Device - The event occurred on a personal device.

IS_REMOTE

Remote - The indication of whether the session is remote.

IS_RENEWAL

Renewal - The indication of whether this is a lease/session renewal event.

IS_SYSTEM

System - The indication of whether the object is part of the operating system.

IS_TRUSTED

Trusted Device - The event occurred on a trusted device.

JA3S_FINGERPRINT

JAS3 Fingerprint - The fingerprint of JAS3 string.

JA3S_STRING

JAS3 String - The JAS3 string.

JA3_FINGERPRINT

JA3 Fingerprint - The fingerprint of JA3 string.

JA3_STRING

JA3 String - The JA3 string.

JOB

Job - The job object that pertains to the event.

KB_ARTICLES

Knowledgebase Articles - The KB article/s related to the entity

KERNEL

Kernel - The kernel resource object that pertains to the event.

KEYBOARD_INFO

Keyboard Information - The keyboard detailed information.

KEYBOARD_LAYOUT

Keyboard Layout - The keyboard locale identifier name (e.g., en-US).

KEYBOARD_SUBTYPE

Keyboard Subtype - The keyboard numeric code.

KEYBOARD_TYPE

Keyboard Type - The keyboard type (e.g., xt, ico).

KEY_LENGTH

Key Length - The length of the encryption key.

LABELS

Labels - The list of labels attached to an event, object, or attribute.

LANG

Language - The two letter lower case language codes, as defined by ISO 639-1. For example: en (English), de (German), or fr (French).

LAST_RUN_TIME

Last Run - The last run time of application or service. See specific usage.

LAST_SEEN_TIME

Last Seen - The most recent detection time of the activity or object. See specific usage.

LATENCY

Latency - TODO: The HTTP response latency. In seconds, milliseconds, etc.?

LEASE_DUR

Lease Duration - This represents the length of the DHCP lease in seconds. This is present in DHCP Ack events. (activity_id = 1)

LENGTH

Response Length - The HTTP response length, in number of bytes.

LICENSE

Software License - The name or identifier of the license applied on package or software. See SPDX License List.

LINEAGE

Lineage - The lineage of the process, represented by a list of paths for each ancestor process. For example: ['/usr/sbin/sshd', '/usr/bin/bash', '/usr/bin/whoami']

LOADED_MODULES

Loaded Modules - The list of loaded module names.

LOAD_TYPE

Load Type - The load type, normalized to the caption of the load_type_id value. In the case of ‘Other’, it is defined by the event source. It describes how the module was loaded in memory.

LOAD_TYPE_ID

Load Type ID - The normalized identifier of the load type. It identifies how the module was loaded in memory.

LOCATION

Detailed Geo Location - The detailed geographical location usually associated with an IP address.

LOGGED_TIME

Logged Time - The time when the logging system collected and logged the event. This attribute is distinct from the event time in that event time typically contain the time extracted from the original event. Most of the time, these two times will be different.

LOGON_PROCESS

Logon Process - The trusted process that validated the authentication credentials.

LOGON_TYPE

Logon Type - The logon type, normalized to the caption of the logon_type_id value. In the case of ‘Other’, it is defined by the event source.

LOGON_TYPE_ID

Logon Type ID - The normalized logon type identifier.

MAC

MAC Address - The Media Access Control (MAC) address that is associated with the network interface.

MALWARE

Malware - The list of malware identified by a finding.

MESSAGE

Message - The description of the event, as defined by the event source.

MESSAGE_UID

Message UID - The email header Message-Id value, as defined by RFC 5322.

METADATA

Metadata - The metadata associated with the event.

METRICS

Metrics - The general purpose metrics associated with the event. See specific usage.

MFA

Multi Factor Authentication - Indicates whether Multi Factor Authentication was used during authentication.

MIME_TYPE

MIME type - The Multipurpose Internet Mail Extensions (MIME) type of the file, if applicable.

MODEL

Model - The peripheral device model.

MODIFIED_TIME

Modified Time - The time when the object was last modified. See specific usage.

MODIFIER

Modifier - The user that last modified the object associated with the event. See specific usage.

MODULE

Module - The module that pertains to the event.

NAME

Name - The name of the entity. See specific usage.

NAMESPACE

Namespace - The namespace is useful in merger or acquisition situations. For example, when similar entities exists that you need to keep separate.

NETWORK_DRIVER

Network Driver - The network driver used by the container. For example, bridge, overlay, host, none, etc.

NETWORK_INTERFACE

Network Interface - The network interface that is associated with the device.

NETWORK_INTERFACES

Network Interfaces - The network interfaces that are associated with the device, one for each MAC address/IP address combination.Note: The first element of the array is the network information that pertains to the event.

NEXT_RUN_TIME

Next Run - The next run time. See specific usage.

OBSERVABLES

Observables - The observables associated with the event.

OPCODE

DNS Opcode - The DNS opcode specifies the type of the query message.

OPCODE_ID

DNS Opcode ID - The DNS opcode ID specifies the normalized query message type.

OPEN_MASK

Open Mask - The Windows options needed to open a registry key.

OPEN_TYPE

Open Type - The file open type.

OPERATION

Operation - Verb/Operation associated with the request

OPNUM

Opnum - An operation number used to identify a specific remote procedure call (RPC) method or a method in an interface.

ORCHESTRATOR

Orchestrator - The orchestrator managing the container, such as ECS, EKS, K8s, OpenShift, None.

ORG_UID

Org ID - The unique identifier of the organization to which the user belongs. For example, Active Directory or AWS Org ID.

ORG_UNIT

Org Unit - The name of the organization to which the user belongs.

ORIGINAL_TIME

Original Time - The original event time as reported by the event source.

OS

OS - The device operation system.

OVERALL_SCORE

Overall Score - The overall score as reported by the event source. See specific usage.

OWNER

Owner - The user that owns the file/object.

PACKAGES

Software Packages - List of vulnerable packages as identified by the security product

PACKETS

Total Packets - The total number of packets (in and out).

PACKETS_IN

Packets In - The number of packets sent from the destination to the source.

PACKETS_OUT

Packets Out - The number of packets sent from the source to the destination.

PACKET_UID

Packet UID - The packet identifier assigned by the protocol.

PARENT_FOLDER

Parent Folder - The parent folder in which the file resides. For example: c:\windows\system32

PARENT_PROCESS

Parent Process - The parent process of this process object. It is recommended to only populate this field for the first process object, to prevent deep nesting.

PATH

Path - The path that pertains to the event or object. See specific usage.

PERIPHERAL_DEVICE

Peripheral Device - The peripheral device that triggered the event.

PERMISSION

Permission - The IAM permission related to an event

PHYSICAL_HEIGHT

Physical Height - The numeric physical height of display.

PHYSICAL_ORIENTATION

Physical Orientation - The numeric physical orientation of display.

PHYSICAL_WIDTH

Physical Width - The numeric physical width of display.

PID

Process ID - The process identifier, as reported by the operating system. Process ID (PID) is a number used by the operating system to uniquely identify an active process.

PIXEL_BITS

Pixel Bits - The number of bits per pixel.

POLICY

Policy - Describes details of a policy. See specific usage.

PORT

Port - The TCP/UDP port number associated with a connection. See specific usage.

POSTAL_CODE

Postal Code - The postal code of the location.

PREFIX

Prefix - The domain prefix.

PREV_REG_KEY

Previous Registry Key - The registry key before the mutation

PREV_REG_VALUE

Previous Registry Value - The registry value before the mutation

PRIORITY

Priority - The priority, normalized to the caption of the priority_id value. In the case of ‘Other’, it is defined by the event source. See specific usage.

PRIORITY_ID

Priority ID - The normalized priority. See specific usage.

PRIVILEGES

Privileges - The user or group privileges.

PROCESS

Process - The process object.

PROCESSED_TIME

Processed Time - The event processed time, such as an ETL operation.

PRODUCT

Product - The product that reported the event.

PRODUCT_UID

Product Identifier - Unique Identifier of a product.

PROFILES

Profiles - The list of profiles used to create the event.

PROJECT_UID

Project ID - The unique identifier of a Cloud project.

PROTOCOL_NAME

Protocol Name - The TCP/IP protocol name in lowercase, as defined by the Internet Assigned Numbers Authority (IANA). See Protocol Numbers. For example: tcp or udp.

PROTOCOL_NUM

Protocol Number - The TCP/IP protocol number, as defined by the Internet Assigned Numbers Authority (IANA). Use -1 if the protocol is not defined by IANA. See Protocol Numbers. For example: 6 for TCP and 17 for UDP.

PROTOCOL_VER

Protocol Version - The Protocol version, normalized to the caption of the protocol_ver_id value. In the case of ‘Other’, it is defined by the event source.

PROTOCOL_VER_ID

Protocol Version ID - The normalized identifier of the Protocol version.

PROVIDER

Provider - The origin of information associated with the event. See specific usage.

PROXY

Proxy - If a proxy connection is present, the connection from the client to the proxy server.

QUERY

DNS Query - The Domain Name System (DNS) query.

QUERY_STRING

HTTP Query String - The query portion of the URL. For example: the query portion of the URL http://www.example.com/search?q=bad&sort=date is q=bad&sort=date.

QUERY_TIME

Query Time - The Domain Name System (DNS) query time.

RAM_SIZE

RAM Size - The ctotal amount of installed RAM, in Megabytes. For example: 2048.

RAW_DATA

Raw Data - The event data as received from the event source.

RAW_HEADER

Raw Header - The email authentication header.

RCODE

Response Code - The server response code, normalized to the caption of the rcode_id value. In the case of ‘Other’, it is defined by the event source.

RCODE_ID

Response Code ID - The normalized identifier of the server response code. See specific usage.

RDATA

DNS RData - The data describing the DNS resource. The meaning of this data depends on the type and class of the resource record.

REFERENCES

References - Supporting reference URLs

REFERRER

HTTP Referrer - The request header that identifies the address of the previous web page, which is linked to the current web page or resource being requested.

REGION

Region - The name or the code of a region. See specific usage.

REGISTRAR

Domain Registrar - The domain registrar.

REG_KEY

Registry Key - The registry key.

REG_VALUE

Registry Value - The registry value.

RELATED_EVENTS

Related Events - Describes events related to a finding or detection as identified by the security product.

RELATED_VULNERABILITIES

Related Vulnerabilities - List of vulnerabilities that are related to this vulnerability.

RELAY

Relay - The network relay that is associated with the event.

RELEASE

Software Release Details - Release is the number of times a version of the software has been packaged.

REMEDIATION

Remediation - The remediation recommendations on how to fix the identified issue(s).

REMOTE_DISPLAY

Remote Display - The remote display affiliated with the event

REPLY_TO

Reply To - The email header Reply-To values, as defined by RFC 5322.

REPUTATION

Reputation Scores - Contains the original and normalized reputation scores.

REQUEST

API Request Details - General Purpose API Request Object. See specific usage

REQUESTED_PERMISSIONS

Requested Permissions - The permissions mask that were requested by the process.

REQUIREMENTS

Requirements - A list of applicable compliance requirements for which this finding is related to.

RESOURCE

Resource - The target resource.

RESOURCES

Resources Array - Describes details about resources that were affected by the activity/event.

RESOURCE_TYPE

Resource Type - The context in which a resource was retrieved in a web request.

RESOURCE_UID

Resource ID - The unique identifier of a cloud resource. For example, S3 Bucket name, EC2 Instance Id.

RESPONSE

API Response Details - General Purpose API Response Object. See specific usage.

RESPONSE_TIME

Response Time - The Domain Name System (DNS) response time.

RISK_LEVEL

Risk Level - The risk level, normalized to the caption of the risk_level_id value. In the case of ‘Other’, it is defined by the event source.

RISK_LEVEL_ID

Risk Level ID - The normalized risk level id.

RISK_SCORE

Risk Score - The risk score as reported by the event source.

RULE

Rule - The rules that reported the events.

RUNTIME

Runtime - The runtime managing this container.

RUN_STATE

Run State - The state of the job or service, normalized to the caption of the run_state_id value. In the case of ‘Other’, it is defined by the event source. See specific usage.

RUN_STATE_ID

Run State ID - The normalized identifier of the state of the job or service. See specific usage.

SAMESITE

SameSite - The cookie attribute that lets servers specify whether/when cookies are sent with cross-site requests. Values are: Strict, Lax or None

SANDBOX

Sandbox - The name of the containment jail (i.e., sandbox). For example, hardened_ps, high_security_ps, oracle_ps, netsvcs_ps, or default_ps.

SANS

Subject Alternative Names - The list of subject alternative names that are secured by a specific certificate.

SCALE_FACTOR

Scale Factor - The numeric scale factor of display.

SCHEME

Scheme - The scheme portion of the URL. For example: http, https, ftp, or sftp.

SCORE

Reputation Score - The reputation score, normalized to the caption of the score_id value. In the case of ‘Other’, it is defined by the event source.

SCORE_ID

Reputation Score ID - The normalized reputation score identifier.

SECURE

Secure - The cookie attribute to only send cookies to the server with an encrypted request over the HTTPS protocol.

SECURITY_DESCRIPTOR

Security Descriptor - The object security descriptor.

SEQUENCE

Sequence Number - Sequence number of the event. The sequence number is a value available in some events, to make the exact ordering of events unambiguous, regardless of the event time precision.

SERIAL_NUMBER

Serial Number - The serial number that pertains to the object. See specific usage.

SERVER_CIPHERS

Server Cipher Suites - The server cipher suites that were exchanged during the TLS handshake negotiation.

SERVER_HASSH

Server HASSH - The Server HASSH fingerprinting object.

SERVICE

Service - The service that pertains to the event.

SESSION

User Session - The authenticated user session.

SEVERITY

Severity - The event severity, normalized to the caption of the severity_id value. In the case of ‘Other’, it is defined by the event source.

SEVERITY_ID

Severity ID - The normalized identifier of the event severity. The normalized severity is a measurement the effort and expense required to manage and resolve an event or incident. Smaller numerical values represent lower impact events, and larger numerical values represent higher impact events.

SHARE

Share - The SMB share name.

SHARE_TYPE

Share Type - The SMB share type.

SIGNATURE

Digital Signature - The digital signature of the file.

SIZE

Size - The size of data, in bytes.

SMTP_FROM

SMTP From - The value of the SMTP MAIL FROM command.

SMTP_HELLO

SMTP Hello - The value of the SMTP HELO or EHLO command.

SMTP_TO

SMTP To - The value of the SMTP envelope RCPT TO command.

SNI

Server Name Indication - The Server Name Indication (SNI) extension sent by the client.

SPF

SPF Status - The Sender Policy Framework (SPF) status of the email.

SP_NAME

OS Service Pack - The name of the latest Service Pack.

SP_VER

OS Service Pack Version - The version number of the latest Service Pack.

SRC_ENDPOINT

Source Endpoint - The network source endpoint.

SRC_URL

Source URL - The URL pointing towards the source of an entity. See specific usage.

START_ADDRESS

Start Address - The start address of the execution.

START_TIME

Start Time - The start time of a time period. See specific usage.

STATE

State - The state of the event or object, normalized to the caption of the state_id value. In the case of ‘Other’, it is defined by the event source. See specific usage.

STATE_ID

State ID - The normalized state ID of the event or object. See specific usage.

STATUS

Status - The event status, normalized to the caption of the status_id value. In the case of ‘Other’, it is defined by the event source.

STATUS_CODE

Status Code - The event status code, as reported by the event source.

For example, in a Windows Failed Authentication event, this would be the value of ‘Failure Code’, e.g. 0x18.

STATUS_DETAIL

Status Details - The status details contains additional information about the event outcome.

STATUS_ID

Status ID - The normalized identifier of the event status.

SUBJECT

Subject - The email header Subject value, as defined by RFC 5322.

SUBJECT_DN

Subject Distinguished Name - The certificate subject distinguished name.

SUBNET

Subnet - The subnet mask.

SUBNET_UID

Subnet UID - The unique identifier of a virtual subnet.

SUPPORTING_DATA

Supporting Data - Additional data supporting a finding as provided by security tool

SVC_NAME

Service Name - The service name in service-to-service connections. For example, AWS VPC logs the pkt-src-aws-service and pkt-dst-aws-service fields identify the connection is coming from or going to an AWS service.

SYSTEM_CALL

System Call - The system call that was invoked.

TACTICS

Tactics - The a list of tactic ID’s/names that are associated with the attack technique, as defined by ATT&CK MatrixTM.

TAG

Image Tag - The image tag. For example: 1.11-alpine.

TCP_FLAGS

TCP Flags - The network connection TCP header flags (i.e., control bits).

TECHNIQUE

Technique - The attack technique.

TERMINATED_TIME

Terminated Time - The time when the entity was terminated. See specific usage.

TEXT

URL Text - The URL. For example: http://www.example.com/download/trouble.exe.

TID

Thread ID - The Identifier of the thread associated with the event, as returned by the operating system.

TIME

Event Time - The normalized event occurrence time.

TIMEZONE_OFFSET

Timezone Offset - The number of minutes that the reported event time is ahead or behind UTC, in the range -1,080 to +1,080.

TITLE

Title - The title of an entity. See specific usage.

TLS

TLS - The Transport Layer Security (TLS) attributes.

TO

To - The email header To values, as defined by RFC 5322.

TRAFFIC

Traffic - The network traffic refers to the amount of data moving across a network at a given point of time. Intended to be used alongside Network Connection.

TRANSACTION_UID

Transaction UID - The unique identifier of the transaction.

TREE_UID

Tree UID - The tree id is a unique SMB identifier which represents an open connection to a share.

TRUE_COLOR

True Color - A boolean indicating whether to extract pixel values through red/green/blue intensities.

TTL

TTL - The time interval that the resource record may be cached. Zero value means that the resource record can only be used for the transaction in progress, and should not be cached.

TYPE

Type - The type of an object or value, normalized to the caption of the type_id value. In the case of ‘Other’, it is defined by the event source. See specific usage.

TYPES

Types - The type/s of an entity. See specific usage.

TYPE_ID

Type ID - The normalized type identifier of an object. See specific usage.

TYPE_NAME

Type Name - The event type name, as defined by the type_uid.

TYPE_UID

Type ID - The event type ID. It identifies the event’s semantics and structure. The value is calculated by the logging system as: class_uid * 100 + activity_id.

UID

Unique ID - The unique identifier. See specific usage.

UNMAPPED

Unmapped Data - The attributes that are not mapped to the event schema. The names and values of those attributes are specific to the event source.

URL

URL - The URL object that pertains to the event or object. See specific usage.

USER

User - The user that pertains to the event or object.

USER_AGENT

HTTP User-Agent - The request header that identifies the operating system and web browser.

USER_RESULT

User Result - The result of the user account change. It should contain the new values of the changed attributes.

UUID

UUID - The universally unique identifier. See specific usage.

VALUE

Value - The value that pertains to the object. See specific usage.

VECTOR_STRING

Vector String - The CVSS vector string is a text representation of a set of CVSS metrics. It is commonly used to record or transfer CVSS metric information in a concise form. For example: 3.1/AV:L/AC:L/PR:L/UI:N/S:U/C:H/I:N/A:H.

VENDOR_NAME

Vendor Name - The name of the vendor. See specific usage.

VERSION

Version - The version that pertains to the event or object. See specific usage.

VLAN_UID

VLAN - The Virtual LAN identifier.

VNC_AUTH

VNC Authentication - The Virtual Network Computing (VNC) authentication object describes the VNC authentication values.

VPC_UID

VPC UID - The unique identifier of the Virtual Private Cloud (VPC).

VULNERABILITIES

Vulnerabilities - This object describes vulnerabilities reported in a security finding.

VULNERABILITY

Vulnerability - The vulnerability object describes details related to the observed vulnerability

WIN_RESOURCE

Windows Resource - The Windows resource object that was accessed, such as a mutant or timer.

XATTRIBUTES

Extended Attributes - An unordered collection of zero or more name/value pairs where each pair represents a file or folder extended attribute. For example: Windows alternate data stream attributes (ADS stream name, ADS size, etc.), user-defined or application-defined attributes, ACL, owner, primary group, etc. Examples from DCS:

X_FORWARDED_FOR

X-Forwarded-For - The X-Forwarded-For header identifying the originating IP address(es) of a client connecting to a web server through an HTTP proxy or a load balancer.

X_ORIGINATING_IP

X-Originating-IP - The X-Originating-IP header identifying the emails originating IP address(es).

ZONE

Network Zone - The network zone or LAN segment.